Optimize loading and caching
