+++ .babelrc
@@ -1,0 +1,6 @@
+{
+  "presets": ["next/babel"],
+  "targets": {
+    "esmodules": true
+  }
+}
