+++ .browserslistrc
@@ -1,0 +1,4 @@
+> 0.5%
+last 2 versions
+not dead
+not ie <= 11
